move zurg tests
