+++ generators/typescript/utils/core-utilities/zurg/babel.config.cjs
@@ -1,1 +1,1 @@
-module.exports = require("../../../babel.config.json");
+module.exports = require("../../../../../babel.config.json");
