+++ generators/typescript/utils/core-utilities/zurg/src/__test__/lazy/lazyObject.test.ts
@@ -1,0 +1,18 @@
+import { lazyObject, number, object, string } from "../../builders";
+import { itSchemaIdentity } from "../utils/itSchema";
+
+describe("lazy", () => {
+    itSchemaIdentity(
+        lazyObject(() => object({ foo: string() })),
+        { foo: "hello" }
+    );
+
+    itSchemaIdentity(
+        lazyObject(() => object({ foo: string() })).extend(object({ bar: number() })),
+        {
+            foo: "hello",
+            bar: 42
+        },
+        { title: "returned schema has object utils" }
+    );
+});
