+++ generators/typescript/utils/core-utilities/zurg/src/__test__/lazy/recursive/a.ts
@@ -1,0 +1,7 @@
+import { object } from "../../../builders/object";
+import { schemaB } from "./b";
+
+// @ts-expect-error
+export const schemaA = object({
+    b: schemaB
+});
