+++ generators/typescript/utils/core-utilities/zurg/src/__test__/lazy/recursive/b.ts
@@ -1,0 +1,8 @@
+import { object } from "../../../builders/object";
+import { optional } from "../../../builders/schema-utils";
+import { schemaA } from "./a";
+
+// @ts-expect-error
+export const schemaB = object({
+    a: optional(schemaA)
+});
