+++ generators/typescript/utils/core-utilities/zurg/src/__test__/object/objectWithoutOptionalProperties.test.ts
@@ -1,0 +1,21 @@
+import { objectWithoutOptionalProperties, string, stringLiteral } from "../../builders";
+import { itSchema } from "../utils/itSchema";
+
+describe("objectWithoutOptionalProperties", () => {
+    itSchema(
+        "all properties are required",
+        objectWithoutOptionalProperties({
+            foo: string(),
+            bar: stringLiteral("bar").optional()
+        }),
+        {
+            raw: {
+                foo: "hello"
+            },
+            // @ts-expect-error
+            parsed: {
+                foo: "hello"
+            }
+        }
+    );
+});
