+++ generators/typescript/utils/core-utilities/zurg/src/__test__/primitives/any.test.ts
@@ -1,0 +1,6 @@
+import { any } from "../../builders";
+import { itSchemaIdentity } from "../utils/itSchema";
+
+describe("any", () => {
+    itSchemaIdentity(any(), true);
+});
