+++ generators/typescript/utils/core-utilities/zurg/src/__test__/primitives/boolean.test.ts
@@ -1,0 +1,14 @@
+import { boolean } from "../../builders";
+import { itSchemaIdentity } from "../utils/itSchema";
+import { itValidate } from "../utils/itValidate";
+
+describe("boolean", () => {
+    itSchemaIdentity(boolean(), true);
+
+    itValidate("non-boolean", boolean(), {}, [
+        {
+            path: [],
+            message: "Expected boolean. Received object."
+        }
+    ]);
+});
