+++ generators/typescript/utils/core-utilities/zurg/src/__test__/primitives/number.test.ts
@@ -1,0 +1,14 @@
+import { number } from "../../builders";
+import { itSchemaIdentity } from "../utils/itSchema";
+import { itValidate } from "../utils/itValidate";
+
+describe("number", () => {
+    itSchemaIdentity(number(), 42);
+
+    itValidate("non-number", number(), "hello", [
+        {
+            path: [],
+            message: 'Expected number. Received "hello".'
+        }
+    ]);
+});
