+++ generators/typescript/utils/core-utilities/zurg/src/__test__/primitives/string.test.ts
@@ -1,0 +1,14 @@
+import { string } from "../../builders";
+import { itSchemaIdentity } from "../utils/itSchema";
+import { itValidate } from "../utils/itValidate";
+
+describe("string", () => {
+    itSchemaIdentity(string(), "hello");
+
+    itValidate("non-string", string(), 42, [
+        {
+            path: [],
+            message: "Expected string. Received 42."
+        }
+    ]);
+});
