+++ generators/typescript/utils/core-utilities/zurg/src/__test__/primitives/unknown.test.ts
@@ -1,0 +1,6 @@
+import { unknown } from "../../builders";
+import { itSchemaIdentity } from "../utils/itSchema";
+
+describe("unknown", () => {
+    itSchemaIdentity(unknown(), true);
+});
