--- generators/typescript/utils/core-utilities/zurg/src/builders/lazy/__test__/recursive/a.ts
@@ -1,7 +1,0 @@
-import { object } from "../../../object";
-import { schemaB } from "./b";
-
-// @ts-expect-error
-export const schemaA = object({
-    b: schemaB
-});
