--- generators/typescript/utils/core-utilities/zurg/src/builders/lazy/__test__/recursive/b.ts
@@ -1,8 +1,0 @@
-import { object } from "../../../object";
-import { optional } from "../../../schema-utils";
-import { schemaA } from "./a";
-
-// @ts-expect-error
-export const schemaB = object({
-    a: optional(schemaA)
-});
