--- generators/typescript/utils/core-utilities/zurg/src/builders/primitives/__test__/any.test.ts
@@ -1,6 +1,0 @@
-import { itSchemaIdentity } from "../../../__test__/utils/itSchema";
-import { any } from "../any";
-
-describe("any", () => {
-    itSchemaIdentity(any(), true);
-});
