--- generators/typescript/utils/core-utilities/zurg/src/builders/primitives/__test__/boolean.test.ts
@@ -1,14 +1,0 @@
-import { itSchemaIdentity } from "../../../__test__/utils/itSchema";
-import { itValidate } from "../../../__test__/utils/itValidate";
-import { boolean } from "../boolean";
-
-describe("boolean", () => {
-    itSchemaIdentity(boolean(), true);
-
-    itValidate("non-boolean", boolean(), {}, [
-        {
-            path: [],
-            message: "Expected boolean. Received object."
-        }
-    ]);
-});
