--- generators/typescript/utils/core-utilities/zurg/src/builders/primitives/__test__/number.test.ts
@@ -1,14 +1,0 @@
-import { itSchemaIdentity } from "../../../__test__/utils/itSchema";
-import { itValidate } from "../../../__test__/utils/itValidate";
-import { number } from "../number";
-
-describe("number", () => {
-    itSchemaIdentity(number(), 42);
-
-    itValidate("non-number", number(), "hello", [
-        {
-            path: [],
-            message: 'Expected number. Received "hello".'
-        }
-    ]);
-});
