--- generators/typescript/utils/core-utilities/zurg/src/builders/primitives/__test__/string.test.ts
@@ -1,14 +1,0 @@
-import { itSchemaIdentity } from "../../../__test__/utils/itSchema";
-import { itValidate } from "../../../__test__/utils/itValidate";
-import { string } from "../string";
-
-describe("string", () => {
-    itSchemaIdentity(string(), "hello");
-
-    itValidate("non-string", string(), 42, [
-        {
-            path: [],
-            message: "Expected string. Received 42."
-        }
-    ]);
-});
