--- generators/typescript/utils/core-utilities/zurg/src/builders/primitives/__test__/unknown.test.ts
@@ -1,6 +1,0 @@
-import { itSchemaIdentity } from "../../../__test__/utils/itSchema";
-import { unknown } from "../unknown";
-
-describe("unknown", () => {
-    itSchemaIdentity(unknown(), true);
-});
