--- generators/typescript/utils/core-utilities/zurg/src/builders/undiscriminated-union/__test__/undiscriminatedUnion.test.ts
@@ -1,46 +1,0 @@
-import { itSchema, itSchemaIdentity } from "../../../__test__/utils/itSchema";
-import { object, property } from "../../object";
-import { number, string } from "../../primitives";
-import { undiscriminatedUnion } from "../undiscriminatedUnion";
-
-describe("undiscriminatedUnion", () => {
-    itSchemaIdentity(undiscriminatedUnion([string(), number()]), "hello world");
-
-    itSchemaIdentity(undiscriminatedUnion([object({ hello: string() }), object({ goodbye: string() })]), {
-        goodbye: "foo"
-    });
-
-    itSchema(
-        "Correctly transforms",
-        undiscriminatedUnion([object({ hello: string() }), object({ helloWorld: property("hello_world", string()) })]),
-        {
-            raw: { hello_world: "foo " },
-            parsed: { helloWorld: "foo " }
-        }
-    );
-
-    it("Returns errors for all variants", async () => {
-        const result = await undiscriminatedUnion([string(), number()]).parse(true);
-        if (result.ok) {
-            throw new Error("Unexpectedly passed validation");
-        }
-        expect(result.errors).toEqual([
-            {
-                message: "[Variant 0] Expected string. Received true.",
-                path: []
-            },
-            {
-                message: "[Variant 1] Expected number. Received true.",
-                path: []
-            }
-        ]);
-    });
-
-    describe("compile", () => {
-        // eslint-disable-next-line jest/expect-expect
-        it("doesn't compile with zero members", () => {
-            // @ts-expect-error
-            () => undiscriminatedUnion([]);
-        });
-    });
-});
